requirements.txt
flask
google-generativeai
sqlalchemy
pytesseract
pillow
pdf2image
python-docx
python-dotenv
pdfplumber